# gen/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "description": "REST API generated from CUE ontology. All endpoints accept/return JSON.",
        "title": "Propeller Property Management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "description": "Local development",
            "url": "http://localhost:8080"
        }
    ],
    "paths": {
        "/v1/persons": {
            "post": {
                "operationId": "CreatePerson",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "summary": "Create a new person",
                "tags": [
                    "Person"
                ]
            },
            "get": {
                "operationId": "ListPersons",
                "parameters": [
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 50,
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "offset",
                        "schema": {
                            "default": 0,
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Person"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "List persons with filtering",
                "tags": [
                    "Person"
                ]
            }
        },
        "/v1/persons/{id}": {
            "get": {
                "operationId": "GetPerson",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Get person by ID",
                "tags": [
                    "Person"
                ]
            },
            "patch": {
                "operationId": "UpdatePerson",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Update person fields",
                "tags": [
                    "Person"
                ]
            }
        },
        "/v1/organizations": {
            "post": {
                "operationId": "CreateOrganization",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrganizationCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organization"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "summary": "Create a new organization",
                "tags": [
                    "Organization"
                ]
            },
            "get": {
                "operationId": "ListOrganizations",
                "parameters": [
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 50,
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "offset",
                        "schema": {
                            "default": 0,
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "List organizations",
                "tags": [
                    "Organization"
                ]
            }
        },
        "/v1/organizations/{id}": {
            "get": {
                "operationId": "GetOrganization",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organization"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Get organization by ID",
                "tags": [
                    "Organization"
                ]
            },
            "patch": {
                "operationId": "UpdateOrganization",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrganizationUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organization"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Update organization",
                "tags": [
                    "Organization"
                ]
            }
        },
        "/v1/person-roles": {
            "post": {
                "operationId": "CreatePersonRole",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonRoleCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonRole"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "summary": "Assign a role to a person",
                "tags": [
                    "PersonRole"
                ]
            },
            "get": {
                "operationId": "ListPersonRoles",
                "parameters": [
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 50,
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "offset",
                        "schema": {
                            "default": 0,
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PersonRole"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "List person roles",
                "tags": [
                    "PersonRole"
                ]
            }
        },
        "/v1/person-roles/{id}": {
            "get": {
                "operationId": "GetPersonRole",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonRole"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Get person role by ID",
                "tags": [
                    "PersonRole"
                ]
            }
        },
        "/v1/person-roles/{id}/activate": {
            "post": {
                "operationId": "ActivateRole",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonRole"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "409": {
                        "description": "Invalid State Transition"
                    }
                },
                "summary": "Activate a pending person role (transitions to \"active\")",
                "tags": [
                    "PersonRole"
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PersonCreate": {
                "type": "object",
                "properties": {
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "date_of_birth": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "ssn_last_four": {
                        "type": "string"
                    },
                    "contact_methods": {
                        "items": {
                            "type": "object"
                        },
                        "type": "array"
                    },
                    "preferred_contact": {
                        "enum": [
                            "email",
                            "sms",
                            "phone",
                            "mail",
                            "portal"
                        ],
                        "type": "string"
                    },
                    "language_preference": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    },
                    "do_not_contact": {
                        "type": "boolean"
                    },
                    "identity_verified": {
                        "type": "boolean"
                    },
                    "verification_method": {
                        "enum": [
                            "manual",
                            "id_check",
                            "credit_check",
                            "ssn_verify"
                        ],
                        "type": "string"
                    },
                    "verified_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "first_name",
                    "last_name",
                    "display_name",
                    "contact_methods",
                    "preferred_contact",
                    "language_preference",
                    "do_not_contact",
                    "identity_verified"
                ]
            },
...
        }
    }
}